refactor(protocol): decompose flist/write/tests.rs into submodules (SPL-17 #2440)
